+++ src/Message/Options.php
@@ -66,6 +66,13 @@ class Options implements Arrayable
     protected $isDryRun = false;
 
     /**
+     * @internal
+     *
+     * @var bool
+     */
+    protected $directBootOk = false;
+
+    /**
      * Options constructor.
      *
      * @param OptionsBuilder $builder
@@ -80,6 +87,7 @@ class Options implements Arrayable
         $this->timeToLive = $builder->getTimeToLive();
         $this->restrictedPackageName = $builder->getRestrictedPackageName();
         $this->isDryRun = $builder->isDryRun();
+        $this->directBootOk = $builder->isDirectBootOk();
     }
 
     /**
@@ -93,6 +101,7 @@ class Options implements Arrayable
         $mutableContent = $this->isMutableContent ? true : null;
         $delayWhileIdle = $this->delayWhileIdle ? true : null;
         $dryRun = $this->isDryRun ? true : null;
+        $directBootOk = $this->directBootOk ? true : null;
 
         $options = [
             'collapse_key' => $this->collapseKey,
@@ -103,6 +112,7 @@ class Options implements Arrayable
             'time_to_live' => $this->timeToLive,
             'restricted_package_name' => $this->restrictedPackageName,
             'dry_run' => $dryRun,
+            'direct_boot_ok' => $directBootOk,
         ];
 
         return array_filter($options);