+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [v1.x.x]
+
+- Add support for setDirectBootOk on OptionsBuilder
+- Mark setDryRun as deprecated v1 FCM on OptionsBuilder
+
 ## [v1.5.0]
 
 - Add support for images on PayloadNotificationBuilder